rename sub components
